♻️ refactor: move TryCatch contract interfaces to `Contracts` namespace
Relocated `ExecutableTryCatchInterface`, `ThrowableHandlerInterface`, and `ThrowableHandlerCollectionInterface` to the `Atournayre\Contracts\TryCatch` namespace. Updated all references accordingly.
